+++ admin/other.php
@@ -2,7 +2,6 @@
 // Direktzugriff auf diese Datei verhindern:
 defined( 'ABSPATH' ) or die();
 	$text = '';
-	//
 	include 'help/fullscreen.php';
 	//
 	include 'help/hidemarkers.php';