+++ extensions-leaflet-map.php
@@ -5,7 +5,7 @@
  * GitHub Plugin URI: https://github.com/hupe13/extensions-leaflet-map-github
  * Primary Branch:    main
  * Description:       Extensions for the WordPress plugin Leaflet Map Github Version
- * Version:           3.1-220707
+ * Version:           3.1-220710
  * Requires PHP:      7.4
  * Author:            hupe13
  * License:           GPL v2 or later
@@ -60,6 +60,8 @@ if (is_admin()) {
   include_once LEAFEXT_PLUGIN_DIR . '/php/providers.php';
   include_once LEAFEXT_PLUGIN_DIR . '/php/safari.php';
 
+  include_once LEAFEXT_PLUGIN_DIR . '/php/leaflet-directory.php';
+  include_once LEAFEXT_PLUGIN_DIR . '/php/managefiles.php';
 
 // Add settings to plugin page
 function leafext_add_action_links ( $actions ) {