+++ changes.md
@@ -1,3 +1,5 @@
 * leaflet-gestures-1.4.3: Please check your Leaflet Map settings for scrollwheel. See the help for Gesture Handling!
 * Leaflet.Control.FullScreen 2.4.0
 * non-admin interface
+* File Manager
+* leaflet-dir: Tracks from all files in a directory